+++ esp-hal-embassy/CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Bump Rust edition to 2024, bump MSRV to 1.85. (#3391)
 - Update `defmt` to 1.0 (#3416)
+- The `IntoAnyTimer` trait has been removed (#3444)
+- The `TimerCollection` trait has been sealed and renamed to `TimeBase`. Former `IntoAnyTimer` functionality has been merged into `TimeBase`. (#3444)
+- `esp_hal_embassy::init` will panic if called multiple times (#3444)
 
 ### Fixed